Experimental addition to the vanilla Recipe Book! This fixes the weird duplication of items upon crafting! The MainMenu will now directly open the RecipeCreator instead of the editor menu. (Recipes can be edited/deleted/disabled via the Recipe List). Fixed the Recipe Book not loading recipes for non existing switch categories.

+++ src/main/java/me/wolfyscript/customcrafting/data/cache/EliteWorkbench.java
@@ -6,7 +6,7 @@ import org.bukkit.inventory.ItemStack;
 
 public class EliteWorkbench {
 
-    private int currentGridSize;
+    private byte currentGridSize;
     private EliteWorkbenchData eliteWorkbench;
     private ItemStack result;
     private ItemStack[] contents;
@@ -25,11 +25,11 @@ public class EliteWorkbench {
         this.contents = contents;
     }
 
-    public int getCurrentGridSize() {
+    public byte getCurrentGridSize() {
         return currentGridSize;
     }
 
-    public void setCurrentGridSize(int currentGridSize) {
+    public void setCurrentGridSize(byte currentGridSize) {
         this.currentGridSize = currentGridSize;
     }
